+++ SafeTrace_App/SafeTrace/package.json
@@ -6,11 +6,25 @@
     "ios": "expo start --ios",
     "web": "expo start --web",
     "eject": "expo eject",
-    "test": "jest --watchAll",
+    "test": "jest --watch --coverage=false --changedSince=origin/master",
+    "testDebug": "jest -o --watch --coverage=false",
+    "testFinal": "jest",
+    "updateSnapshots": "jest -u --coverage=false",
     "lint": "tslint --project tsconfig.json"
   },
   "jest": {
-    "preset": "jest-expo"
+    "preset": "jest-expo",
+    "transformIgnorePatterns": [
+      "node_modules/(?!(jest-)?react-native|react-clone-referenced-element|@react-native-community|expo(nent)?|@expo(nent)?/.*|react-navigation|@react-navigation/.*|@unimodules/.*|unimodules|sentry-expo|native-base|@sentry/.*)"
+    ],
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "json",
+      "jsx",
+      "node"
+    ]
   },
   "dependencies": {
     "@expo/vector-icons": "~10.0.6",
@@ -44,6 +58,7 @@
     "rxjs": "^6.5.5",
     "seamless-immutable": "^7.1.4",
     "styled-components": "^5.1.0",
+    "ts-jest": "^25.3.1",
     "tslint": "^6.1.1",
     "tslint-config-prettier": "^1.18.0",
     "tslint-eslint-rules": "^5.4.0",
@@ -51,14 +66,17 @@
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
+    "@types/jest": "^25.2.1",
     "@types/ramda": "^0.27.3",
     "@types/react": "^16.9.33",
     "@types/react-native": "^0.62.2",
     "@types/react-redux": "^7.1.7",
+    "@types/react-test-renderer": "^16.9.2",
     "@types/seamless-immutable": "^7.1.11",
     "@types/styled-components": "^5.0.1",
     "babel-preset-expo": "~8.1.0",
     "jest-expo": "~37.0.0",
+    "react-test-renderer": "^16.13.1",
     "reactotron-react-native": "^5.0.0",
     "redux-tools": "^3.2.0",
     "typescript": "^3.8.3"
