+++ SafeTrace_App/SafeTrace/tsconfig.jest.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig",
+  "compilerOptions": {
+    "jsx": "react",
+    "module": "CommonJS"
+  }
+}
